[ownership] Emit a PrettyStackTrace msg telling to re-run with -sil-verify-all if we hit a verifier error before ownership lowering.
We have been getting a bunch of different bugs reported that were initially
identified as being caused by the ownership model eliminator (due to this
verification failure), but that turned out in reality to be a different failure.

To ease triaging, this commit puts in a pretty stack trace msg that tells the
person doing the triaging to re-run with -sil-verify-all enabled to track down
the actual pass that introduces the verification failure.