[windows] Adapt Windows build script to latest changes.
LLDB in Windows was failing to build in the community CI. From comparing
the invocation of build-script-impl and windows-build.bat it seems that
we have migrated from LLDB_PATH_TO_*_SOURCE/BUILD variables to *_DIR
variables.

Additionally I needed to disable building the test on Windows because
the community CI doesn't build LLD, which seems to be needed for the
tests. I don't understand exactly why before the test were being skipped
automatically.

Also, a lot of CMake variables have been annotated with their expected
types. At least the new *_DIR variables were being interpreted as
strings in one occassion and the path separators were making CMake fail.
The other case was a boolean that must have been interpreted as a
string.