[sil] Rename SILBasicBlock::getParent() -> getFunction().
This is a refactoring that we have talked about for a while with an eye towards
providing a consistent name across all IR constructs for getting its
corresponding SILFunction (if it has one).

This refactoring was done using Xcode's refactoring engine. I then
git-clang-formatted the result and eliminated a couple of places where
clang-format didn't do a good job with ternary operations.