[AutoDiff]: Typo in TransposingAttr::TransposingAttr
Previously, TransposingAttr::TransposingAttr would set the DeclAttribute to be
DAK_Differentiating. This is incorrect, as it should be DAK_Transposing. This
commit fixes that.

Note: I noticed that this bug hasn't been observed because this code path is
currently unused. However, it is planned to be used in the future, so I will
not remove the code. This unfortunately makes it impossible to write a test
that would have caught this regression.