[Diagnostics] Move macros used in diagnostic definition files to their own header
These macros were becoming very complicated and hard to maintain since they were
duplicated across a dozen-or-so files. Move them into a header which can be
imported as needed.

Added the INTERNALERROR and INTERNALWARNING macros which expand to ERROR and
WARNING respectively. These exist only to document diags which have been
audited for public status but do not meet the criteria.