[mandatory-combine] Simplify block traversal used to initialize worklist.
Previously, we were using a pointer set to ensure that was checked when a value
was popped off the worklist. In this commit, I change the code so that we
instead use the set to guard insertion into the worklist.

I changed the name of the set to blockAlreadyAddedToWorklist. Hopefully that
will make things a bit clearer.