Put More Faith In Raw Values From Synthesized Enums
By requestifying the raw value accessor and pushing the semantic check
into IRGen, we are now partially validating enums that are
synthesized by the clang importer and deserialization - something that
we were not doing before.  The previous check would only skip
the auto-incrementer step when this was the case.  We now also skip the
uniqueness check.

Resolves rdar://56037318