[AutoDiff] [SIL] Tweak 'differentiable_function' syntax.
In `differentiable_function` instruction's syntax, use `parameters` instead of `wrt` and 'with_derivative' instead of `with`. This is to align with `linear_function` instruction and the future direction that both parameter indices and result indices will be included in this instruction.

Note: `with_derivative` is not named `with_derivatives` because VJPs will be dropped from this instruction when we complete JVP + linear map transposition.

Resolves [TF-909](https://bugs.swift.org/browse/TF-909).