[AutoDiff] Destroy all pullback indirect results after adjoint accumulation.
When we differentiate a function (example below) with respect to a proper subset of its indirect parameters and when the function only has a derivative with respecct to a superset of those indirect parameters, the pullback returns more indirect results that what we need. However, unneeded indirect results are not destroyed, which causes a memory lifetime verification failure. This patch fixes this bug by releasing all pullback indirect results instead of just releasing the ones needed for calculating the derivative.

Resolves [TF-914](https://bugs.swift.org/browse/TF-914).