[NFC] Refactor Override Fixit Emission
Refactor diagnostic emission so it lazily emits notes and fixits.
This is a necessary evil in a world where an arbitrary request can also
emit diagnostics, especially circularity diagnostics.  In the future, we
should assert in the evaluator that there is no active diagnostic to
catch the rest of these.