[gardening] Eliminate unneeded llvm:: namespace from containers in allocbox to stack.
The specific types are:

* SmallVector{,Impl}
* SmallPtrSet{,Impl}
* StringRef

At the time AllocBoxToStack was not written, these containers had not been
injected into the swift namespace. They are now, so this commit just eliminates
the unnecessary llvm:: namespace in favor of referring to these containers
without qualification.

I also changed a SmallSet to be a SmallPtrSet. SmallSet's implementation just
delegates to SmallPtrSet anyways, so this just eliminates an anti-pattern.