Use 80-character column width.
Originally, the doc used 80-character column width.

When cross-posting the doc to the forums, we found that Discourse Markdown
treats newlines as actual line breaks, so we removed all newlines:
https://forums.swift.org/t/differentiable-programming-mega-proposal/28547

Since the forum post is up, I think we can make the doc be the source-of-truth
with 80-character column width.