SILCombine: fix a problem with thin_to_thick_function elimination.
The optimizer can generate a thin_to_thick_function conversion from a throwing thin to a non-throwing thick function (in case it can prove that the function is actually not throwing).
Therefore, when removing such a  conversion from the callee-argument of an apply, we have to check if the new callee (= the argument of the thin_to_thick_function) is a throwing function and set the not-throwing flag in this case.

This fixes a SILVerifier crash.
rdar://problem/56358645