[ownership] When RAUWing multiple value instruction results, insert destroys if we move up the lattice from .owned -> .none.
Consider the following SIL:

%83 = enum $Optional<Thread>, #Optional.none!enumelt // users: %93, %84
%84 = tuple (%82 : $Result<Array<UInt8>, AnyError>, %83 : $Optional<Thread>) // user: %85
(%85, %86) = destructure_tuple %84 : $(Result<Array<UInt8>, AnyError>, Optional<Thread>) // user: %91

In this case, we clearly can replace all uses of %86 with %83. But since %86 is
a multiple value instruction result, despite our RAUWing, we can not eliminate
it. This implies since we have erased %83 being .none when passing it as an
argument to %84, we need to insert a destroy.

This will eventually be a no-op in the pipeline when the destructure is torn
apart by later passes.

In order to prevent constant propagation from introducing an infinite loop due
to the insertion of these destroys, I included a set in the constant folder to
ensure that we do not revisit results that have only constant folder inserted
destroy uses.