Mark loaded_module_trace_swiftinterface as unsupported in Windows
Sadly, in Windows, the last invocation of the test, which mixes
`emit-loaded-module-trace` and the two frontend invocations, fails with
the compiler crashing. Instead of this being reported correctly thru
lit, it crashes the lit driver and the test is not even marked as
failed. When running several tests in parallel, it crashes one of the
workers and outputs a cryptic "Error 87".

Sadly, because of how the test seems to break, and how Windows seems to
fail, XFAILing the test is not enough, and it has to be skipped
completely in Windows and have to be mark as unsupported instead.

Hopefully this will unstuck the Windows CI machines.